backend/requirements.txt
Django>=6.0
djangorestframework
dj-database-url
python-dotenv
django-cors-headers
requests
groq
